Add Mux.Context
Bind modules' values in the scope of a function using the process
dictionary. The intention is for Mux.ServerSession and Mux.ClientSession
to automatically broadcast the bound values when appropriate, and to
provide helpers for functions where it would be cumbersome to carry the
state around (similar to :rand).

The binding API is very explicit to try to avoid confusion. Likely this
will only be directly called by Mux.