Only handle Mux.Trace by default
Mux.Trace would only be implicitly present when better tracer is
not used because it would need to be explicitly added on client
side and explicitly overridden on server side.

Mux.Deadline could cause nack without being obvious and could read
the time when not required, i.e. immediately after setting it.